requirements.txt
Flask==3.1.3
flask-cors==6.0.2
flask-login==0.6.3
werkzeug==3.1.7
python-dotenv==1.2.2
mysql-connector-python==9.6.0
mysqlclient==2.2.8
